chore(senna): update gulp-metal dependency to latest (2.2.2)
As I said in the previous commit, even if I try updating gulp-metal from
the one we were using (1.12.2) to the latest (2.2.2) in order to get a
newer version of node-sass (eg. 4.14.1), I still get a duplicate old one
(4.5.3) nested under node_modules/gulp-metal/node_modules/node-sass/.

It took me a long time to figure this out, but the reason is for an npm
feature I didn't even remember existed, and which we are abusing
horribly here, to ship basically the entire dependency graph of
gulp-metal as an invisible, unoverridable blob:

    https://stackoverflow.com/questions/11207638/advantages-of-bundleddependencies-over-normal-dependencies-in-npm

Yep, we really ship all of these packages *inside* gulp-metal:

    https://github.com/metal/gulp-metal/blob/17218b70ace3f0a733b6f49ef95d581c1e5c26bf/package.json#L73-L103

Observe how a `wget https://registry.npmjs.org/gulp-metal/-/gulp-metal-2.2.2.tgz`
has us downloading a 50 MB archive:

    Resolving registry.npmjs.org (registry.npmjs.org)... 104.16.22.35, 104.16.25.35, 104.16.26.35, ...
    Connecting to registry.npmjs.org (registry.npmjs.org)|104.16.22.35|:443... connected.
    HTTP request sent, awaiting response... 200 OK
    Length: 52343983 (50M) [application/x-tar]

So this is why `yarn why node-sass` swears there is only one version on
the disk:

    => Found "node-sass@4.14.1"
    info Reasons this module exists
       - "gulp-metal#gulp-sass" depends on it
       - Hoisted from "gulp-metal#gulp-sass#node-sass"

And why `yarn why gulp-sass` says there's one version of that:

    => Found "gulp-sass@3.2.1"
    info Reasons this module exists
       - "gulp-metal" depends on it
       - Hoisted from "gulp-metal#gulp-sass"

When in reality we have:

- node_modules/gulp-sass at 3.2.1, asking for node-sass@^4.8.3
- node_modules/node-sass at 4.14.1
- node_modules/gulp-metal "asking for" gulp-sass@^3.1.0
- node_modules/gulp-metal/node_modules/gulp-sass at 3.1.0, "asking for"
  node-sass@^4.2.0
- node_modules/gulp-metal/node_modules/node-sass at 4.5.3

In addition to being invisible to `yarn why`, these hidden nested
versions cannot be overridden with a `"resolutions"` field in the
package.json.

So I'm putting this up here for educational purposes only; the only way
to fix this for real will be to cut a gulp-metal release without this
insanity (will have to wait for Chema to come back for that), or to copy
the project source here and modify it before using it inline, or to rip
out and replace the whole Metal-based build toolchain (sadly, that would
be expensive because it is complicated and that means we'd have to pull
it apart carefully to figure out which bits of it actually matter; I
think Senna's build requirements are actually pretty simple underneath
all the layers of abstraction and packages like `metal-tools-build-*`
and so on, but it would take time to figure that out).